Add #elif to enable an external mz_crc32() to be linked in.
Adds '#elif defined(USE_EXTERNAL_MZCRC)' by the mz_crc32() definition
to enable the option of a user linking in an alternate crc implementation.

The main reason why this might be desired would be to use an SSE-accelerated
crc implementaiton, which would be faster, but not be reasonable to include
in this file.